fix another compile error on high boost

+++ src/opentrade/algo.cc
@@ -387,7 +387,7 @@ void AlgoManager::SetTimeout(Algo::IdType id, std::function<void()> func,
   kTimers.emplace(kTime + seconds * 1e6, func);
 #else
   auto t = new boost::asio::deadline_timer(
-      io_service_, boost::posix_time::microseconds(seconds * 1e6));
+      io_service_, boost::posix_time::microseconds((int64_t)(seconds * 1e6)));
   t->async_wait(strands_[id % threads_.size()].wrap([func, t](auto) {
     func();
     delete t;
